Prepare receiving of OutputMessages for using fibers
A client first registers via DataStorage::RegisterForOutputMessage that
it wants to receive an OutputMessage and gets a boost::fibers::future
back.  As soon as a matching message arrives, the DataStorage sets the
value of the future.

If an OutputMessage arrives such that no client has registered a need
for it (RegisterForOutputMessage was not called), it is dropped.
Furthermore only one client can register for a OutputMessage at a time.