Update Conditions for gate setup/online counters
Introduce an additional counter for the number of gate setup routines
executed.  Also, since the counters themselves are atomic, we do not
need to lock a mutex for incrementing.  We should only call NotifyAll
when all setup/online phases are evaluated and not after each increment.
Moreover, make getter methods const.