Add flags for completion of gate setup/online
The register maintains two counters which are atomically incremented
whenever the setup or online phase, respectively, for a gate is
completed.  The main thread is supposed to wait via condition variables
until the counters have reached the total number of gates in the
circuit.

In the previous implementation, a deadlock could occur when the final
increment of the counter and the following notify on the condition
variable happened after the main thread checked the condition but before
it started waiting on the condition variable. I.e.,

Main:   * Acquires the mutex corresponding to the condition variable
        * Checks the condition (not yet satisfied)
Worker: * Atomically increments den counter (without acquiring the mutex)
        * Condition is fulfilled
        * Calls notify_all on the condition variable
Main:   * Starts waiting on the condition variable

This patch introduces Boolean flags which are protected by the mutex
corresponding to the condition variable.  They are set when the counters
have reached their final value and the main thread is able to wait by
using the same mutex with the condition variable.  This solves the
potential deadlock described above while avoiding having to acquire a
mutex upon each increment.