Join fibers created in the FiberThreadPool
The fibers created by the FiberThreadPool need to be joined before
terminating the worker threads.  Otherwise, it might be tried to let
fibers run on a thread after the thread's scheduler has been shutdown.

In order to prevent repeated allocations to store the fiber objects, the
expected number of tasks is given as argument to the constructor of the
FiberThreadPool.