Allocate smaller stacks for worker fibers
Use a custom allocator for the fiber stack.  14K stack space seems to be
sufficient for know, and is a significant improvement over the default
stack size of 128K.  Also, add constants to configure the stack size and
switch to protected stacks, respectively.  The latter adds guard pages
at the end of each stack.