Rewrite communication subsystem
The high-level interface for the application is `communication_layer`.
It provides methods to send and broadcast messages to other parties, and
allows to register instances of `message_handler` for incoming messages
of a certain type.  The low-level send and receive functionality is
provided by classes implementing the `transport` interface.
`tcp_transport` implements message passing over a TCP stream using
Boost.Asio.  `dummy_transport` is intended for local testing purposes
and based on a pair of synchronized queues.