Use unique instead of shared_ptr for OT objects
The OTProvider does not need to access the *OT*Sender/Receiver objects.
Thus, the client gets exclusive ownership of the created object via a
unique_ptr.  This also removes the overhead of the reference-counting
shared_ptr.

Functions which destroy a unique_ptr object with the default deleter,
e.g., destructors, require a complete type, a forward declaration is not
sufficient.  Hence, such function definitions are moved to the .cpp
file.