replace MessageHandler with MessageManager
MessageManager replaces CommunicationHandler to simplify the API that, as I noticed, was hard to understand and use for unfamiliar developers.
MessageHandler requires creating new datastructures and adding new logic to the code that manages how message arrive, where they are stored, etc. So, a lot of boilerplate code that needs to be learned by heart.
MessageManager, on the other hand, stores a data structure to register and access future/promise based communication channels. Namely, a future is registered via the MessageManager API and an arriving message gets forwarded to promises[sender_id][message_type][message_id].
I didn't notice significant differences in the running time. Tests in "Release" run from about as fast as before to about 15% faster. May be just fluctuations.